backend/requirements.txt
fastapi
uvicorn
python-dotenv
supabase
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pytest
httpx
sentry-sdk
python-jose